Se implemnto el Login

+++ SADVO/Controllers/CiudadanoController.cs
@@ -10,15 +10,24 @@ namespace SADVO.Controllers
     {
 
         private readonly ICiudadanoService _ciudadanoService;
+        private readonly IUsuarioSession _usuarioSession;
 
 
-        public CiudadanoController(ICiudadanoService ciudadanoService)
+        public CiudadanoController(ICiudadanoService ciudadanoService, IUsuarioSession usuarioSession)
         {
             _ciudadanoService = ciudadanoService;
+            _usuarioSession = usuarioSession;
         }
 
         public async Task<IActionResult> Index()
         {
+
+            if (!_usuarioSession.HasUser()) 
+            {
+
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+
+            }
 
             var dtos = await _ciudadanoService.GetAll();
 
@@ -41,51 +50,46 @@ namespace SADVO.Controllers
 
         public IActionResult Create()
         {
+            if (!_usuarioSession.HasUser())
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
 
-            return View("Save", new CiudadanoSaveViewModel() { Id = 0, Nombre = "", Apellido = "", Email = "", DocumentoIdentidad = "", EstaActivo = false  });
-
+            return View("Save", new CiudadanoSaveViewModel() { Id = 0, Nombre = "", Apellido = "", Email = "", DocumentoIdentidad = "", EstaActivo = false });
         }
 
         [HttpPost]
         public async Task<IActionResult> Create(CiudadanoSaveViewModel vm)
         {
+            if (!_usuarioSession.HasUser())
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
 
             if (!ModelState.IsValid)
-            {
                 return View("Save", vm);
-            }
 
-            CiudadanoDto dto = new() {
+            CiudadanoDto dto = new()
+            {
                 Id = 0,
                 Nombre = vm.Nombre,
                 Apellido = vm.Apellido,
                 Email = vm.Email,
                 DocumentoIdentidad = vm.DocumentoIdentidad,
-                EstaActivo = vm.EstaActivo };
+                EstaActivo = vm.EstaActivo
+            };
 
             await _ciudadanoService.AddAsync(dto);
             return RedirectToRoute(new { controller = "Ciudadano", action = "Index" });
-
-
         }
-
 
         public async Task<IActionResult> Delete(int Id)
         {
+            if (!_usuarioSession.HasUser())
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
 
             if (!ModelState.IsValid)
-            {
                 return RedirectToRoute(new { controller = "Ciudadano", action = "Index" });
-            }
 
             var dto = await _ciudadanoService.GetById(Id);
             if (dto == null)
-            {
-
                 return RedirectToRoute(new { controller = "Ciudadano", action = "Index" });
-
-            }
-
 
             CiudadanoDeleteViewModel vm = new()
             {
@@ -98,69 +102,58 @@ namespace SADVO.Controllers
             };
 
             return View(vm);
-
-
         }
-
 
         [HttpPost]
         public async Task<IActionResult> DeleteConfirmacion(int id)
         {
+            if (!_usuarioSession.HasUser())
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
+
             var result = await _ciudadanoService.DeleteAsync(id);
 
             if (!result)
-            {
+                TempData["Error"] = "El ciudadano no pudo ser eliminado";
+            else
+                TempData["Succes"] = "El ciudadano fue eliminado correctamente";
 
-                TempData["Error"] = "El ciudadano no pudo ser elimiando";
-
-            }
-            else 
-            {
-                TempData["Succes"] = "El ciudadano no pudo ser eliminado";
-            }
-
-            return RedirectToAction ("Index");
-
-
-
-
-
+            return RedirectToAction("Index");
         }
-
-
 
         public async Task<IActionResult> Edit(int Id)
         {
+            if (!_usuarioSession.HasUser())
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
 
             if (!ModelState.IsValid)
-            {
                 return RedirectToRoute(new { controller = "Ciudadano", action = "Index" });
-            }
 
             ViewBag.EditMode = true;
             var dto = await _ciudadanoService.GetById(Id);
             if (dto == null)
-            {
-
                 return RedirectToRoute(new { controller = "Ciudadano", action = "Index" });
 
-            }
-
-            CiudadanoSaveViewModel vm = new() { Id = dto.Id, Nombre = dto.Nombre, Apellido = dto.Apellido, Email = dto.Email, DocumentoIdentidad = dto.DocumentoIdentidad, EstaActivo = dto.EstaActivo };
+            CiudadanoSaveViewModel vm = new()
+            {
+                Id = dto.Id,
+                Nombre = dto.Nombre,
+                Apellido = dto.Apellido,
+                Email = dto.Email,
+                DocumentoIdentidad = dto.DocumentoIdentidad,
+                EstaActivo = dto.EstaActivo
+            };
 
             return View("Save", vm);
-
         }
-
 
         [HttpPost]
         public async Task<IActionResult> Edit(CiudadanoSaveViewModel vm)
         {
+            if (!_usuarioSession.HasUser())
+                return RedirectToRoute(new { controller = "Login", action = "Index" });
 
             if (!ModelState.IsValid)
-            {
                 return View("Save", vm);
-            }
 
             CiudadanoDto dto = new()
             {
@@ -175,5 +168,6 @@ namespace SADVO.Controllers
             await _ciudadanoService.UpdateAsync(dto);
             return RedirectToRoute(new { controller = "Ciudadano", action = "Index" });
         }
+
     }
 }
